problem with getPositionID i think
