+++ FootballManager/app/views/fixtures.scala.html
@@ -110,7 +110,7 @@
         <!-- add a submit button -->
         <div class="actions">
 
-            @helper.form(action = routes.Application.upload) {
+            @helper.form(action = routes.LeagueCtrl.upload) {
             <p>
                 <input class="btn-sm btn-primary" type="submit" value="Generate Fixtures">
             </p>
