+++ FootballManager/app/views/register.scala.html
@@ -61,7 +61,7 @@
                                         <!-- Add a submit button -->
                                     <div class="actions">
                                     <input type="submit" value="Submit"  class="btn btn-primary">
-                                        <a href="@routes.Application.register()">
+                                        <a href="@routes.RegisterCtrl.register()">
                                         <button class="btn btn-warning">Cancel</button>
                                         </a>
